рефакторинг: перенос файлов, вынесение таймера в отдельный файл
